Refactored the package name for OSS
