+++ src/main/java/io/boomerang/Application.java
@@ -1,4 +1,4 @@
-package net.boomerangplatform;
+package io.boomerang;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
