+++ src/main/java/io/boomerang/SecurityConfig.java
@@ -1,3 +1,4 @@
+package io.boomerang;
 //package net.boomerangplatform;
 //
 //import org.apache.logging.log4j.LogManager;
