+++ src/main/java/io/boomerang/SecurityFilter.java
@@ -1,3 +1,4 @@
+package io.boomerang;
 //package net.boomerangplatform;
 //
 //import java.io.IOException;
